refactor and add ErrorPage
