+++ client/src/components/index.ts
@@ -4,6 +4,7 @@ export { HttpCatPage } from './HttpCatPage'
 export { RandomDogPage } from './RandomDogPage'
 export { ClientsRegistrationPage } from './ClientsRegistrationPage'
 export { RegisterPage } from './RegisterPage'
+export { NotFoundPage } from './NotFoundPage'
 
 export { Header } from './Header'
 export { Pagination } from './Pagination'
